spring-cloud-dataflow-common-security/src/main/java/org/springframework/cloud/dataflow/common/security/support/ConditionalOnIdentityProvider.java
/*
 * Copyright 2020 the original author or authors.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      https://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package org.springframework.cloud.dataflow.common.security.support;

import java.lang.annotation.Documented;
import java.lang.annotation.ElementType;
import java.lang.annotation.Retention;
import java.lang.annotation.RetentionPolicy;
import java.lang.annotation.Target;

import org.springframework.context.annotation.Conditional;

/**
 * {@link Conditional @Conditional} that checks if a spesific identity provider is configured.
 *
 * @author Janne Valkealahti
 */
@Target({ ElementType.TYPE, ElementType.METHOD })
@Retention(RetentionPolicy.RUNTIME)
@Documented
@Conditional(OnIdentityProviderCondition.class)
public @interface ConditionalOnIdentityProvider {

	Type type() default Type.NONE;

	enum Type {

		/**
		 * Effectively no security enabled as no identity provider or security
		 * properties set
		 */
		NONE,

		/** Legacy type where old-style oauth2 settings are used */
		LEGACY,

		/** Azure AD type */
		AZUREAD
	}
}
